backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis (cache)
redis==5.0.1

# Validação e serialização
pydantic==2.5.0
pydantic-settings==2.1.0

# Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilitários
python-dateutil==2.8.2
pytz==2023.3

# Upload de arquivos
openpyxl==3.1.2
pandas==2.1.4

# HTTP client (para testes)
httpx==0.25.2

# Desenvolvimento
pytest==7.4.3
pytest-asyncio==0.21.1

email-validator>=2.0.0